feat: add interface `IReadOnlyBox<T>`
This interface will be used to represent a box that may or may not
contain a value.

By accepting this interface (e.g. as a parameter to a method), it
implies that you will not change the contents of the box (although you
may mutate the value itself if it is mutable).

Implementations of this interface may still change the value inside the
box.

Resolves: #5